refactor(error): use composition over typealias
